refactor(EncryptedCardInfo): move to another aggregate
- 암호화 자체가 꾀나 무거운 책임이기 때문에 분리 필요
- 카드 정보 암호화 관련 책임을 별도의 aggregate로 이관
