+++ src/main/java/com/kakao/preinterview/payment/domain/encrypt/EncryptDecrypt.java
@@ -1,4 +1,4 @@
-package com.kakao.preinterview.payment.utils;
+package com.kakao.preinterview.payment.domain.encrypt;
 
 import javax.crypto.Cipher;
 import javax.crypto.SecretKey;
