+++ src/main/java/com/kakao/preinterview/payment/domain/encrypt/EncryptedCardInfo.java
@@ -1,6 +1,8 @@
-package com.kakao.preinterview.payment.domain.payment;
+package com.kakao.preinterview.payment.domain.encrypt;
 
-public class EncryptedCardInfo {
+import com.kakao.preinterview.payment.domain.payment.CardInfo;
+
+public class EncryptedCardInfo implements EncryptedValue {
     private String encryptedValue;
 
     protected EncryptedCardInfo(String encryptedValue) {
