+++ src/main/java/com/kakao/preinterview/payment/domain/encrypt/EncryptedValue.java
@@ -1,0 +1,5 @@
+package com.kakao.preinterview.payment.domain.encrypt;
+
+public interface EncryptedValue {
+    String getEncryptedValue();
+}
