+++ src/test/java/com/kakao/preinterview/payment/domain/cardcompany/CardCompanyInfoTests.java
@@ -1,6 +1,8 @@
 package com.kakao.preinterview.payment.domain.cardcompany;
 
 import com.kakao.preinterview.payment.domain.cardcompany.exceptions.ParseToCardCompanyDataException;
+import com.kakao.preinterview.payment.domain.encrypt.EncryptedValue;
+import com.kakao.preinterview.payment.domain.encrypt.FakeEncryptedValue;
 import com.kakao.preinterview.payment.domain.payment.FakePaymentInfoFactory;
 import com.kakao.preinterview.payment.domain.payment.Payment;
 import org.junit.jupiter.api.BeforeEach;
@@ -13,23 +15,25 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 class CardCompanyInfoTests {
     private Payment fakePayment;
     private String paymentSampleStringData = "_446PAYMENT___XXXXXXXXXXXXXXXXXXXX1234567890123456____001125777____1100000000010000____________________YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY_______________________________________________________________________________________________________________________________________________________________________________________________________________________________________________________".replaceAll("_", " ");
+    private EncryptedValue encryptedValue;
 
     @BeforeEach
     public void setup() {
         fakePayment = FakePaymentInfoFactory.createFakePayment();
+        encryptedValue = new FakeEncryptedValue();
     }
 
     @DisplayName("현재 Payment 객체의 상태를 카드사 전송용 형태로 바꿀 수 있어야 함")
     @Test
     void parseToCardCompanyTest() {
-        CardCompanyInfo cardCompanyInfo = CardCompanyInfo.createCardCompanyInfo(fakePayment);
+        CardCompanyInfo cardCompanyInfo = CardCompanyInfo.createCardCompanyInfo(fakePayment, encryptedValue);
         assertThat(cardCompanyInfo.getStringData()).isEqualTo(paymentSampleStringData);
     }
 
     @DisplayName("현재 Payment 객체 상태를 전송용 데이터로 변경했는데 길이가 안 맞을 경우 error")
     @Test
     void parseToCardCompanyValidationTest() {
-        CardCompanyInfo cardCompanyInfo = CardCompanyInfo.createCardCompanyInfo(fakePayment);
+        CardCompanyInfo cardCompanyInfo = CardCompanyInfo.createCardCompanyInfo(fakePayment, encryptedValue);
         assertThatThrownBy(() -> cardCompanyInfo.validateStringDataLength("hello"))
                 .isInstanceOf(ParseToCardCompanyDataException.class);
     }
