+++ src/test/java/com/kakao/preinterview/payment/domain/encrypt/FakeEncryptedValue.java
@@ -1,0 +1,8 @@
+package com.kakao.preinterview.payment.domain.encrypt;
+
+public class FakeEncryptedValue implements EncryptedValue {
+    @Override
+    public String getEncryptedValue() {
+        return "YYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYYY";
+    }
+}
